How These Thresholds Work MOVING_DISTANCE_THRESHOLD (50 meters):
This is the minimum distance the user must move before the app stores a new location.

If the user moves less than 50 meters, the app will not store the location.

STABLE_TIME_THRESHOLD (60 seconds):

This is the maximum time interval between location updates when the user is stationary.

If the user does not move for 60 seconds, the app will store the location.